+++ FastAnimations/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added support for slingshot animations.
+
 ## 1.10.2
 Released 05 June 2022 for SMAPI 3.14.0 or later.